Breaking Record apart from Reader
Abstract namedtuple allows to just have a simple representation
of any abstract. getattr and get allow for safe access to properties,
they just return None if the item doesn't exist rather then
raising an exception.

Pulling abstract into a list comprehension allows us to just say 'we have
abstracts, abstracts have these properties', and we don't have to worry to
test if they do have this or that property in our XML. We also have a single
return type now, a list, rather than a possible None, a list with a single
string, or a list of dicts with different possible keys.

Classification is even simpler, with just pulling out the text. Both
classification and abstract will return an empty list if they don't exist.

Setting methods as @property allows to just call record.abstract rather
than record.abstract(). Not a huge savings but it looks a bit cleaner

MODSReader class is a simplistic reader. It just creates an iterator from an XML file and returns the next element until it's exhausted. Bare bones.

Serializing the tree to a string in the __next__ of MODSREADER
and turning it back into a etree in Record is somewhat inefficient
but allows us to just open single-record XML files as a Record
rather than making a MODSReader and getting the first record.