This is not a working implementation.
There is a mixture of working and broken code in here, so
don't try to run it as is. There are places where I thought
having the original with comments would be better than
rewriting the code. To run the code or the tests, comment out
the subject and rights methods, it should go then. This is just
an easier way of me giving you some feedback that I hope is helpful.

The overarching theme is that we can get rid of a lot of existence
checks and simplify code. Also, look for places where you can just
pull the information in a field rather than checking around for what it is.

Generally, decoupling is hard but worthwhile. Having different code
that is responsible for different parts makes for reusable and
extendable software. So look for places where you've baked assumptions
or workflows into the library and consider whether you should just
extract the information with this tool and do something interesting with
it with another.

Bottom bottom line, great job, automating stuff like this is really
hard and you did a bang-up job! Cheers!